requirements.txt
discord.py>=2.5.0
fastapi>=0.111.0
uvicorn>=0.29.0
httpx>=0.27.0
jinja2>=3.1.3
pydantic>=2.8.0
python-dotenv>=1.0.1
aiosqlite>=0.20.0
